requirements.txt
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
et-xmlfile==1.1.0
layout==0.4.1
numpy==1.21.2
openpyxl==3.0.10
pandas==1.3.2
pdfminer.six==20220524
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sortedcontainers==2.4.0
xlsxwriter==3.0.2
PyPDF2==3.0.1